+++ src/main/java/fr/openent/competences/helpers/ExportEvaluationHelper.java
@@ -30,7 +30,7 @@ public class ExportEvaluationHelper {
         devoirMap.put(ID_KEY, devoir.getLong(ID_KEY));
         devoirMap.put("nom", devoir.getString("name"));
         devoirMap.put("coeff", devoir.getString("coefficient"));
-        devoirMap.put("sur", devoir.getLong("diviseur"));
+        devoirMap.put("sur", devoir.getDouble("diviseur"));
         devoirMap.put("periode", I18n.getInstance().translate("viescolaire.periode." + devoir.getLong("periodetype").toString(), language,host)
                 + " " + String.valueOf(devoir.getLong("periodeordre")));
         devoirMap.put("sousMatiere", devoir.getString("libelle", ""));
